requirements.txt
Pillow==9.4.0
pynput==1.7.6
pytesseract==0.3.10
neovim==0.3.1
